requirements.txt
pyTelegramBotAPI==4.14.0
peewee==3.17.0
requests==2.31.0
python-dotenv==1.0.0
python-telegram-bot-calendar==1.0.5
